Fixed epochTime, moved function that connected with run table to run.service.ts, moved function that connected with log table to log.service.ts, removed function getDate, fixed typo error
