+++ src/controllers/createpdf.controller.ts
@@ -5,50 +5,53 @@ import { CreateLogInfo } from '../dtos/createpdf.loginfo.dto';
 import { CreateRunConf } from '../dtos/createpdf.runconf.dto';
 import { CreatePdfService } from '../services/createpdf.service';
 import { ApiUseTags } from '@nestjs/swagger';
+import { LogService } from '../services/log.service';
+import { RunService } from '../services/run.service';
 
 @ApiUseTags('createpdf')
 @Controller('createpdf')
 export class CreatepdfController {
-  constructor(public createPdfService: CreatePdfService) {}
+  constructor(public createPdfService: CreatePdfService, private logService: LogService,
+              private runService: RunService) {}
 
   /**
-   * Get Log information in pdf by id. /createpdf/getlog
+   * Get Log information in pdf by id. /createpdf/getlog?id=number
    * @param query there must be one id
    * @param res response to client
    */
   @Get('getlog')
   // query is needed, with body or param doesn't works
   async getLogInfo(@Query() query: CreateLogInfo, @Res() res: Response): Promise<void> {
-    const epochTime = new Date();
+    const epochTime = new (Date)().getTime();
     let params;
     // checking does id exists
     try {
-      params = await this.createPdfService.getLogInfoParams(query.id);
+      params = await this.logService.getLogInfoParams(query.id);
     } catch (err) {
       return res.end('No such ID!');
     }
     return this.createPdfService
-    .createPdf(params, `Log_${query.id}_${epochTime.getTime()}.pdf`, '/templates/LogInfoTemplate.ejs', res);
+    .createPdf(params, `Log_${query.id}_${epochTime}.pdf`, '/templates/LogInfoTemplate.ejs', res);
   }
   /**
    * Get differense between two runs in pdf
-   * by their id. /createpdf/getrunconf
+   * by their id. /createpdf/getrunconf?id1=number&id2=number
    * @param query there must be two id of runs
    * @param res response to client
    */
   @Get('getrunconf')
   // query is needed, with body or param doesn't works
   async runconf(@Query() query: CreateRunConf, @Res() res: Response): Promise<void> {
-    const epochTime = new Date();
+    const epochTime = new (Date)().getTime();
     let params;
     // checking does id exists
     try {
-      params = await this.createPdfService.getRunConfParams(query.id1, query.id2);
+      params = await this.runService.getRunConfParams(query.id1, query.id2);
     } catch (err) {
       return res.end('No such ID!');
     }
     return this.createPdfService
-    .createPdf(params, `Run_${query.id1}_and_${query.id2}_${epochTime.getTime()}`, '/templates/RunTemplate.ejs', res);
+    .createPdf(params, `Run_${query.id1}_and_${query.id2}_${epochTime}`, '/templates/RunTemplate.ejs', res);
   }
   /**
    * Post bug report in pdf from html form
@@ -59,8 +62,8 @@ export class CreatepdfController {
   @Post('downloadPDF')
   makePdf(@Res() res: Response, @Body() userData: CreateUserData): Promise<void> {
 
-    const allDate = this.createPdfService.getDate();
-    const epochTime = new Date();
+    const allDate = new (Date)().toLocaleString('en-GB');
+    const epochTime = new (Date)().getTime();
     const fname = userData.fname;
     const lname = userData.lname;
     const occupationData = userData.occupation;
@@ -69,7 +72,7 @@ export class CreatepdfController {
     const params = {firstname : fname, lastname : lname, occupation : occupationData, bugReport : bugReportData,
       date : allDate};
     return this.createPdfService
-    .createPdf(params, `BugReport_${epochTime.getTime()}.pdf`, '/templates/BugReportTemplate.ejs', res);
+    .createPdf(params, `BugReport_${epochTime}.pdf`, '/templates/BugReportTemplate.ejs', res);
    }
    /**
     * Get html form for bug report. /createpdf
